Adding sender and receiver code
Sender code sends a pulse every 25ms for 11 seconds when the button is pressed via esp-now, this pulse contains information for either red or off depending on the state of the led on the sender when pressed. The receiver wakes up for ~100ms then goes into deep sleep for 10 seconds to conserve the battery of the system and allow for ~4 week runtime on a 5000mah battery.